requirements.txt
dataclasses-json==0.5.7
requests==2.31.0
iris-pex-embedded-python>=3.0.0
iris-embedded-python-wrapper
pytest
requests
intersystems-irispython
notebook
ipykernel
sqlalchemy-iris